lab2/requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.9
neo4j==5.28.1
python-dotenv==1.0.1